// src/types/generated/contract.ts
export interface BindEntry {
  action: string;
  "allow-inhibiting"?: boolean | null;
  "allow-when-locked"?: boolean | null;
  args?: string[] | null;
  "cooldown-ms"?: number | null;
  key: string;
  "overlay-title"?: string | null;
}
// ...
export interface LayoutConfig {
  "always-center-single-column"?: boolean | null;
  "background-color"?: string | null;
  border?: BorderConfig | null;
  "center-focused-column"?: string | null;
  "default-column-display"?: string | null;
  "default-column-width"?: number | null;
  "empty-workspace-above-first"?: boolean | null;
  "focus-ring"?: FocusRingConfig | null;
  gaps?: number | null;
  "preset-column-widths"?: number[] | null;
  "preset-window-heights"?: number[] | null;
  shadow?: ShadowConfig | null;
  struts?: StrutsConfig | null;
  "tab-indicator"?: TabIndicatorConfig | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "BorderConfig".
 */
export interface BorderConfig {
  color?: string | null;
  enabled?: boolean | null;
  gradient?: string | null;
  width?: number | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "FocusRingConfig".
 */
export interface FocusRingConfig {
  color?: string | null;
  enabled?: boolean | null;
  gradient?: string | null;
  width?: number | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "ShadowConfig".
 */
export interface ShadowConfig {
  color?: string | null;
  "draw-behind-window"?: boolean | null;
  enabled?: boolean | null;
  "offset-x"?: number | null;
  "offset-y"?: number | null;
  softness?: number | null;
  spread?: number | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "StrutsConfig".
 */
export interface StrutsConfig {
  bottom?: number | null;
  left?: number | null;
  right?: number | null;
  top?: number | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "TabIndicatorConfig".
 */
export interface TabIndicatorConfig {
  colors?: string[] | null;
  "corner-radius"?: number | null;
  gaps?: number | null;
  position?: string | null;
}
// ...
export interface NamedWorkspace {
  layout?: LayoutConfig | null;
  name: string;
  "open-on-output"?: string | null;
}